ai-trading-system/requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
pytest
pytest-cov
requests
prometheus-client
flake8
tenacity
pytest-mock
python-dotenv
alpaca-trade-api
streamlit
yfinance